Duplicate listing for same model
I've removed multiple listing for the same model. Previously for each model trained on CASIA WebFace and MS1MV2, there were 5 listings named acc. to the test set, which doesn't make sense because the model would remain the same. I've also checked the hashes of the files and they were the same.